Update assignment2.cpp
line 72 is in // (remove it in comment will crash the program - segmentation fault)

Explanation: The conversion from infix to postfix calls the inToPost() method in which I am creating a variable named postfix which assigned the postfix expression. The problem with this is that the class attribute postfix remains empty and in the evaluate() the stack is empty and returning its top creates segmentation fault. To avoid this I simply // comment the line in the inToPost() method where the postfix variable is declared. So, the class attribute will get assigned postfix string.